+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
 - Added "Set Up renv" button to initialize the environment, install
   dependencies, and create a lockfile when no valid R lockfile file is present.
+
+### Fixed
+
+- Fixed an issue where the preferred Python path was incorrectly not found when deploying #2742
 
 ## [1.21.0]